Allowing multiple "statements" inside the if branches.
"Statements" here are in quotes because they are actually not
statements from the parser's point of view. Parser statements
include function declarations, and I didn't want to allow a
function declaration inside an if.  I wanted only top level functions.

So instead of using statements, I changed the definition of
a block to allow only expressions and variable declarations,
and I changed the if to use blocks as branches.